pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "streamagent"
version = "0.1.0"
description = "Streaming agentic inference with persistent KV cache"
requires-python = ">=3.10"
dependencies = [
    "llama-cpp-python>=0.2.90",
    "numpy>=1.26",
    "pydantic>=2.0",
]

[project.optional-dependencies]
hf = [
    "torch>=2.2",
    "transformers>=4.40",
    "accelerate>=0.28",
]
mlx = [
    "mlx>=0.12",
    "mlx-lm>=0.10",
]
eval = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "alfworld>=0.3",
    "pandas>=2.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["streamagent*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=streamagent --cov-report=term-missing --cov-fail-under=80"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true
